% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ENCODE_eCLIP_EIF4G2_K562_narrowPeak.bed}
\alias{ENCODE_eCLIP_EIF4G2_K562_narrowPeak.bed}
\title{ENCODE eCLIP EIF4G2 peak coordinates}
\format{
A tab-separated BED6+4 narrowPeak format file.
}
\source{
\url{https://www.encodeproject.org/experiments/ENCSR307YIW/}
}
\value{
No value is returned. The dataset is stored as a raw file under
  \code{inst/extdata}.
}
\description{
A raw dataset containing eCLIP peak coordinates for EIF4G2 in K562 cells,
retrieved from the ENCODE project portal (Accession: ENCSR307YIW).
}
\references{
Luo, Y., Hitz, B.C., Gabdank, I., Hilton, J.A., Kagda, M.S.,
  Lam, B., Myers, Z., Sud, P., Jou, J., Lin, K., et al. (2020). New developments
  on the Encyclopedia of DNA Elements (ENCODE) data portal. Nucleic Acids Res.
  48, D882–D889.
}
