% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS.R
\name{formatMS}
\alias{formatMS}
\title{Format Mutational Sensitivity Scores into a Matrix}
\usage{
formatMS(
  snv_sensitivity_scores,
  reference_motif,
  reference_score,
  K,
  nucleotides = c("A", "C", "G", "T"),
  sensitivity_method = "1_minus_norm_score"
)
}
\arguments{
\item{snv_sensitivity_scores}{Named numeric vector/list of sensitivity scores
(output from `calSNVSens`).}

\item{reference_motif}{Character string, the original motif.}

\item{K}{Integer, the length of the motif.}

\item{nucleotides}{Character vector of possible nucleotides.}
}
\value{
A matrix (rows=nucleotides, cols=positions 1 to K) with sensitivity scores.
  The cell corresponding to the original nucleotide at a position might be NA or 0.
}
\description{
Internal helper to arrange SNV sensitivity scores into the
standard Nucleotide x Position matrix format.
}
\keyword{internal}
