% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generateBkgSet}
\alias{generateBkgSet}
\title{Generate a Single Set of Background Sequences}
\usage{
generateBkgSet(
  peak_gr,
  genome_obj,
  K,
  bkg_min_dist,
  bkg_max_dist,
  scramble = TRUE
)
}
\arguments{
\item{peak_gr}{A GRanges object representing the original genomic peaks.}

\item{genome_obj}{A BSgenome object from which sequences will be extracted.}

\item{K}{Integer, the K-mer length. This is used to set a minimum length
for sequences extracted by the internal call to `getSequence`.}

\item{bkg_min_dist}{Integer, the minimum absolute distance (in base pairs) by which
peaks should be shifted. Passed as `X` to `genRNDist()`.}

\item{bkg_max_dist}{Integer, the maximum absolute distance (in base pairs)
for shifting peaks. Passed as `max_dist` to `genRNDist()`.}

\item{scramble}{Logical, if TRUE (default), scrambles the background sequences
after extraction. Set to FALSE for faster execution when scrambling is not needed.}
}
\value{
A DNAStringSet object of background sequences (scrambled or not).
  Returns an empty DNAStringSet if no valid background regions/sequences
  could be generated.
}
\description{
Creates one set of background sequences by randomly
  shifting original peak locations, removing overlaps with original peaks,
  and extracting sequences. Optionally scrambles the sequences.
}
\keyword{internal}
