% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{genMotifVar}
\alias{genMotifVar}
\title{Generate Single Nucleotide Variants for a Motif}
\usage{
genMotifVar(motif, type = "DNA")
}
\arguments{
\item{motif}{Character string, the reference motif. Should only contain valid nucleotides.}

\item{type}{Character string specifying the nucleic acid type for the inputmotifand the generated variants. Accepted values are "DNA" (A,C,G,T) or "RNA" (A,C,G,U) (default: 'DNA').}
}
\value{
A character vector of all unique SNV motifs (excluding the original motif).
}
\description{
Create all possible single nucleotide variants.
}
\keyword{internal}
