% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS.R
\name{returnSensitivityAll}
\alias{returnSensitivityAll}
\title{Calculate Sensitivity for All Motifs}
\usage{
returnSensitivityAll(motif_enrichment)
}
\arguments{
\item{motif_enrichment}{A data frame with 'MOTIF' and 'Score' columns.}
}
\description{
Calculate Sensitivity for All Motifs
}
\keyword{internal}
