% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VS.R
\name{returnSensitivitySpecific}
\alias{returnSensitivitySpecific}
\title{Calculate Sensitivity for a Specific Motif}
\usage{
returnSensitivitySpecific(
  motif_enrichment,
  motif,
  output_type,
  sensitivity_method
)
}
\arguments{
\item{motif_enrichment}{A data frame with 'MOTIF' and 'Score' columns.}

\item{motif}{Character string, the reference motif. Only used if
`return_type = "specific"`. If NULL or "top", the top-scoring motif is used.}

\item{output_type}{Character string, specifying the output format if
`return_type = "specific"`: "matrix" (default) or "number". Ignored if
`return_type = "all"`.}

\item{sensitivity_method}{Character string defining how sensitivity is calculated.
Default: "1_minus_norm_score".}
}
\description{
Calculate Sensitivity for a Specific Motif
}
\keyword{internal}
