% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scrambleDNASet}
\alias{scrambleDNASet}
\title{Scramble a Set of DNA Sequences}
\usage{
scrambleDNASet(dna_set)
}
\arguments{
\item{dna_set}{A DNAStringSet object.}
}
\value{
A DNAStringSet object with each sequence independently shuffled.
  Returns an empty DNAStringSet if input is empty.
}
\description{
Vectorized version of \code{\link{scrambleDNA}} that operates on
  an entire DNAStringSet. More efficient than applying scrambleDNA element-wise
  because it avoids N individual DNAString constructor calls, constructing one
  DNAStringSet from a character vector at the end.
}
\keyword{internal}
