% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{valInputMotif}
\alias{valInputMotif}
\title{Validate Motif Input}
\usage{
valInputMotif(motif, kmer_size, available_motifs)
}
\arguments{
\item{motif}{Character string, the motif to validate.}

\item{kmer_size}{Integer, the expected K-mer length.}

\item{available_motifs}{Character vector of all motifs present in the data.}
}
\value{
Returns `TRUE` invisibly if valid, otherwise stops with an error.
}
\description{
Checks if a user-provided motif is valid (correct length, exists
  within a provided set of available motifs). Stops execution if invalid.
}
\keyword{internal}
