% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{countKmers}
\alias{countKmers}
\title{Count K-mers in Sequences}
\usage{
countKmers(sequences, K, nucleotides = c("A", "C", "G", "T"))
}
\arguments{
\item{sequences}{A DNAStringSet object.}

\item{K}{Integer, the K-mer size.}

\item{nucleotides}{Character vector of nucleotides (default: DNA).}
}
\value{
A data frame with 'MOTIF' and 'COUNT' columns.
}
\description{
Counts occurrences of all possible K-mers within a DNAStringSet.
}
\keyword{internal}
