% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generateBkgSet}
\alias{generateBkgSet}
\title{Generate a Single Set of Background Sequences}
\usage{
generateBkgSet(peak_gr, genome_obj, K, Bkg_dist, max_shift_dist)
}
\arguments{
\item{peak_gr}{A GRanges object representing the original genomic peaks.}

\item{genome_obj}{A BSgenome object from which sequences will be extracted.}

\item{K}{Integer, the K-mer length. This is used to set a minimum length
for sequences extracted by the internal call to `getSequence`.}

\item{Bkg_dist}{Integer, the minimum absolute distance (in base pairs) by which
peaks should be shifted. Passed as `X` to `genRNDist()`.}

\item{max_shift_dist}{Integer, the maximum absolute distance (in base pairs)
for shifting peaks. Passed as `max_dist` to `genRNDist()`.}
}
\value{
A DNAStringSet object of one set of scrambled background sequences.
  Returns an empty DNAStringSet if no valid background regions/sequences
  could be generated.
}
\description{
Creates one set of scrambled background sequences by randomly
  shifting original peak locations, removing overlaps with original peaks,
  extracting sequences, and then scrambling them.
}
\keyword{internal}
