% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getSequence}
\alias{getSequence}
\title{Get Sequences from GRanges}
\usage{
getSequence(granges_obj, genome_obj)
}
\arguments{
\item{granges_obj}{A GRanges object.}

\item{genome_obj}{A BSgenome object.}
}
\value{
A DNAStringSet object containing the sequences.
}
\description{
Extracts DNA sequences for a set of genomic ranges from a BSgenome object.
}
\keyword{internal}
