% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBPSpecificity-package.R
\docType{package}
\name{RBPSpecificity-package}
\alias{RBPSpecificity-package}
\alias{RBPSpecificity}
\alias{#}
\alias{Ensures}
\alias{?RBPSpecificity}
\alias{and}
\alias{?RBPSpecificity-package}
\alias{work}
\title{RBPSpecificity: Analyze RBP Binding Specificity and Mutational Sensitivity}
\description{
The `RBPSpecificity` package provides a suite of tools for researchers
working with RNA Binding Proteins (RBPs) to analyze binding specificity from
high-throughput sequencing data such as eCLIP (enhanced CrossLinking and
ImmunoPrecipitation) and RBNS (RNA Bind-n-Seq).

It allows users to quantify K-mer enrichment within RBP binding sites,
assess the Inherent Specificity (IS) of binding motifs, and evaluate the
Mutational Sensitivity (MS) of these motifs to single nucleotide variations.
The package aims to streamline these common bioinformatic analyses and
provide clear visualizations for interpreting RBP binding characteristics.

**Note:** This package is currently under active development.
}
\section{Key Functions}{

The main workflow of the package revolves around the following exported functions:
\itemize{
  \item \code{\link{motifEnrichment}}: Calculates K-mer enrichment from peak data...
  \item \code{\link{returnIS}}: Determines the Inherent Specificity (IS) score...
  \item \code{\link{plotIS}}: Visualizes the distribution of motif enrichment scores...
  \item \code{\link{returnMS}}: Quantifies the Mutational Sensitivity (MS) of a motif...
  \item \code{\link{plotMS}}: Generates plots to visualize Mutational Sensitivity profiles...
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/S00NYI/RBPSpecificity} (Development repository)
  \item Report bugs at \url{https://github.com/S00NYI/RBPSpecificity/issues}
}
}
\author{
Soon Yi \email{cu.soonyi@gmail.com}
}
